+++ infra_mcdn/imageopt_mcdn.tf
@@ -24,20 +24,17 @@ resource "google_cloud_run_v2_service" "imageopt_svc" {
 }
 
 # CloudRun IAM
-resource "google_cloud_run_v2_service_iam_binding" "imageopt_svc_unauthenticated_invoke" {
+resource "google_cloud_run_v2_service_iam_member" "imageopt_svc_unauthenticated_invoke" {
   location = google_cloud_run_v2_service.imageopt_svc.location
   name     = google_cloud_run_v2_service.imageopt_svc.name
   role     = "roles/run.invoker"
-  members = [
-    "allUsers"
-  ]
+  member   = "allUsers"
 }
 
 # HTTP Load Balancer
 
 # Serverless Network Endpoint Group 
 resource "google_compute_region_network_endpoint_group" "imageopt_svc_neg" {
-  provider              = google
   name                  = "imageopt-svc-neg"
   network_endpoint_type = "SERVERLESS"
   region                = var.cloudrun_region
@@ -61,8 +58,7 @@ resource "google_compute_backend_service" "imageopt_serverless_backend" {
 
 # reserved LB IP address
 resource "google_compute_global_address" "imageopt-lb-ip" {
-  provider = google
-  name     = "imageopt-lb-static-ip"
+  name = "imageopt-lb-static-ip"
 }
 
 # LB url map
@@ -88,7 +84,6 @@ resource "google_compute_target_http_proxy" "imageopt_http_proxy" {
 
 # forwarding rule
 resource "google_compute_global_forwarding_rule" "imageopt_global_fwding_rule" {
-  provider              = google
   name                  = "imageopt-global-rule"
   target                = google_compute_target_http_proxy.imageopt_http_proxy.id
   port_range            = "80"