+++ infra_mcdn/variables.tf
@@ -16,7 +16,7 @@ variable "cloudrun_region" {
 
 variable "imageopt_svc_image" {
   type        = string
-  default     = "us-east1-docker.pkg.dev/img-optimization/img-optimization-repo-1088347617355/image-optimizer:v1.0.6"
+  default     = "us-east1-docker.pkg.dev/img-optimization/img-optimization-repo-1088347617355/image-optimizer:v1.0.7"
   description = "Container image URL in Artifact Registry"
 }